+++ android/src/main/kotlin/dev/note11/flutter_naver_map/flutter_naver_map/model/map/NaverMapViewOptions.kt
@@ -54,6 +54,9 @@ internal data class NaverMapViewOptions(
             val options = NaverMapOptions().apply {
                 compassEnabled(false)
                 zoomControlEnabled(false)
+                scaleBarEnabled(false)
+                logoClickEnabled(false)
+                locationButtonEnabled(false)
             }
 
             val applier = NaverMapOptionApplierImpl(options)