+++ android/src/main/kotlin/dev/note11/flutter_naver_map/flutter_naver_map/view/NaverMapView.kt
@@ -17,7 +17,6 @@ import dev.note11.flutter_naver_map.flutter_naver_map.controller.overlay.Overlay
 import dev.note11.flutter_naver_map.flutter_naver_map.converter.DefaultTypeConverter.asNullableMap
 import dev.note11.flutter_naver_map.flutter_naver_map.model.base.NPoint
 import dev.note11.flutter_naver_map.flutter_naver_map.model.map.NaverMapViewOptions
-import dev.note11.flutter_naver_map.flutter_naver_map.util.NLocationSource
 import dev.note11.flutter_naver_map.flutter_naver_map.util.TextureSurfaceViewUtil
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.platform.PlatformView
@@ -65,8 +64,7 @@ internal class NaverMapView(
 
     private fun onMapReady() {
         initializeMapController()
-        setLocationSource()
-        setMapTapListener()
+        setMapEventListeners()
 
         mapView.onCreate(null)
         naverMapControlSender.onMapReady()
@@ -86,16 +84,14 @@ internal class NaverMapView(
         }
     }
 
-    private fun setLocationSource() {
-        naverMap.locationSource = NLocationSource(activity)
-    }
-
-    private fun setMapTapListener() {
+    private fun setMapEventListeners() {
         isListenerRegistered = true
-
         naverMap.run {
             setOnMapClickListener { pointFPx, latLng ->
                 naverMapControlSender.onMapTapped(NPoint.fromPointFWithPx(pointFPx), latLng)
+            }
+            setOnMapLongClickListener { pointFPx, latLng ->
+                naverMapControlSender.onMapLongTapped(NPoint.fromPointFWithPx(pointFPx), latLng)
             }
             setOnSymbolClickListener {
                 naverMapControlSender.onSymbolTapped(it)
@@ -104,6 +100,7 @@ internal class NaverMapView(
             addOnCameraChangeListener(naverMapControlSender::onCameraChange)
             addOnCameraIdleListener(naverMapControlSender::onCameraIdle)
             addOnIndoorSelectionChangeListener(naverMapControlSender::onSelectedIndoorChanged)
+            addOnLoadListener(naverMapControlSender::onMapLoaded)
 
             naverMap.setCustomStyleId(
                 naverMapViewOptions.naverMapOptions.customStyleId,
@@ -112,14 +109,16 @@ internal class NaverMapView(
         }
     }
 
-    private fun removeMapTapListener() {
+    private fun removeMapEventListeners() {
         if (isListenerRegistered) {
             naverMap.run {
                 onMapClickListener = null
+                onMapLongClickListener = null
                 onSymbolClickListener = null
                 removeOnCameraChangeListener(naverMapControlSender::onCameraChange)
                 removeOnCameraIdleListener(naverMapControlSender::onCameraIdle)
                 removeOnIndoorSelectionChangeListener(naverMapControlSender::onSelectedIndoorChanged)
+                removeOnLoadListener(naverMapControlSender::onMapLoaded)
             }
         }
     }
@@ -128,7 +127,7 @@ internal class NaverMapView(
 
     override fun dispose() {
         unRegisterLifecycleCallback()
-        removeMapTapListener()
+        removeMapEventListeners()
 
         mapView.run {
             onPause()
@@ -187,6 +186,7 @@ internal class NaverMapView(
 
     override fun onConfigurationChanged(newConfig: Configuration) {}
 
+    @Deprecated("Deprecated in Java")
     override fun onLowMemory() {
         mapView.onLowMemory()
     }