+++ lib/src/widget/elements/naver_logo_widget.dart
@@ -12,36 +12,37 @@ class NMapLogoWidget extends StatelessWidget {
     required this.logoClickEnable,
   });
 
+  static const width = 48.0;
+  static const height = 17.0;
+
   @override
   Widget build(BuildContext context) {
     final borderRadius = BorderRadius.circular(99);
     return DecoratedBox(
-      decoration: BoxDecoration(
-          borderRadius: borderRadius,
-          border: Border.all(color: const Color(0x17000000))),
-      child: InkWell(
-          onTap: logoClickEnable
-              ? () => showDialog(
-                  context: context,
-                  builder: (context) =>
-                      NMapInfoDialog(naverMapController: naverMapController))
-              : null,
-          borderRadius: borderRadius,
-          child: Padding(
-            padding: const EdgeInsets.all(1),
-            child: SizedBox(
-                width: 46,
-                height: 15,
+        decoration: BoxDecoration(
+            borderRadius: borderRadius,
+            border: Border.all(color: const Color(0x17000000))),
+        child: SizedBox(
+            width: width,
+            height: height,
+            child: Padding(
+                padding: const EdgeInsets.all(1),
                 child: Material(
                     color: Colors.white,
+                    clipBehavior: Clip.hardEdge,
                     borderRadius: borderRadius,
-                    child: Center(
-                      child: CustomPaint(
-                          size: const Size(35, 7.2),
-                          painter: _NaverLogoPainter()),
-                    ))),
-          )),
-    );
+                    child: InkWell(
+                        onTap: logoClickEnable
+                            ? () => showDialog(
+                                context: context,
+                                builder: (context) => NMapInfoDialog(
+                                    naverMapController: naverMapController))
+                            : null,
+                        child: Center(child: _naverLogo()))))));
+  }
+
+  Widget _naverLogo() {
+    return CustomPaint(size: const Size(35, 7.2), painter: _NaverLogoPainter());
   }
 }