+++ CHANGELOG.md
@@ -1,3 +1,45 @@
+## 1.4.0-dev.2
+
+### Update
+- [All Platform] **Update Naver Map SDK to 3.22.0** ([PR #329](https://github.com/note11g/flutter_naver_map/pull/329))
+
+### Add (New Features)
+- [All Platform] add new api `iconAlpha`, `subIconAlpha`, `setIconAlpha`, `setSubIconAlpha` on `NLocationOverlay`
+
+### Fix
+- [All Platform] Fix NLocale.countryCode support is missing (Fix [#325](https://github.com/note11g/flutter_naver_map/issues/325), [PR #328](https://github.com/note11g/flutter_naver_map/pull/328))
+
+
+## 1.4.0-dev.1
+
+### Breaking Change
+- [All Platform, Not Working Impact] **Migrated built-in location tracking to be widget-based.** ([PR #322](https://github.com/note11g/flutter_naver_map/pull/322))
+  - The `locationButtonEnable` option in `NaverMapViewOptions` has been removed and replaced with the `NMyLocationButtonWidget`.
+  - Location tracking APIs are now provided through the `NMyLocationTracker`.
+  - NaverMapController.setLocationTrackingMode(NLocationTrackingMode)'s return type change from `Future<void>` to `void`
+  - `Future<NLocationTrackingMode> NaverMapController.getLocationTrackingMode()` api exposing change to `NLocationTrackingMode NaverMapController.locationTrackingMode` getter
+  - change `color.alpha` on `NLocationOverlay.defaultCircleColor` 24% to 8%
+- [All Platform, Not Working Impact] **Migrated ScaleBar to be a widget.** ([Fix #297](https://github.com/note11g/flutter_naver_map/issues/297), [PR #319](https://github.com/note11g/flutter_naver_map/pull/319))
+  - The `scaleBarEnable` option in `NaverMapViewOptions` has been removed and replaced with the `NMapScaleBarWidget`.
+- [All Platform] The `showControls` argument in `NaverMapController.takeSnapshot` has been removed as it is no longer supported by the native SDK since version 1.4.0.
+
+### Add (New Features)
+- [All Platform] **Added a new built-in location tracking API, `NDefaultMyLocationTracker`.**
+  - This provides the `NMyLocationTracker` interface, allowing users to implement custom location tracking logic.
+- [All Platform] Added `NaverMap.onMapLoaded` and `NaverMap.onMapLongTapped` callbacks. ([PR #315](https://github.com/note11g/flutter_naver_map/pull/315))
+- [All Platform] Added `NPolygonOverlay.outlinePattern` feature from Naver Map SDK 3.21.0. ([PR #317](https://github.com/note11g/flutter_naver_map/pull/317))
+- [iOS] Added support for Swift Package Manager (SPM). ([Resolve #281](https://github.com/note11g/flutter_naver_map/issues/281), [PR #320](https://github.com/note11g/flutter_naver_map/pull/320))
+
+### Update
+- [All Platform] Improved `NOverlayImage.fromWidget` to automatically detect the widget's size with `null` value. ([Resolve #304](https://github.com/note11g/flutter_naver_map/issues/304))
+- [All Platform] Improved `NLocationTrackingMode.noFollow` to track the current location's bearing. ([PR #323](https://github.com/note11g/flutter_naver_map/pull/323))
+
+### Fix
+- [iOS] Fixed an issue where `contentsPadding` was not applied to the scale bar. ([Fix #297](https://github.com/note11g/flutter_naver_map/issues/297), [PR #319](https://github.com/note11g/flutter_naver_map/pull/319))
+- [iOS] Updated the InfoWindow UI to match the Android version. ([Fix #229](https://github.com/note11g/flutter_naver_map/issues/229))
+- [All Platform] Fixed a `LateInitializationError` that occurred when `NChannelWrapper` was disposed before initialization. ([Fix #312](https://github.com/note11g/flutter_naver_map/issues/312))
+
+
 ## 1.3.1
 - [All Platform, Update] change naver map sdk version to 3.21.0
 - [All Platform, Add] **Add new auth api `FlutterNaverMap().init`** ([#291](https://github.com/note11g/flutter_naver_map/issues/291)), Add detail exception classes. (with sealed class)